refactor(controller, compete-result): 패키지 이동
